1. Unprefix principal urls in getPrincipal method for AccessUtil
2. Changed type of content length from int to long - affects nodes and interfaces. Needed for resources.
    Also means a schema change, stored length needs to be a bigint rather than int. Also content must be stored as a blob.
    Schema was broken anyway, resources limited to 255 bytes. Not being used other than by tests so schema change should be OK.

3. CalDavBwIntf canPut method needs to take account of resource nodes.

4. Null changes to CalDAVCalNode, CalDavComponentNode - finals and @override

5. For resources use streams rather than byte arrays. Avoids copying.

+++ server/src/org/bedework/caldav/server/CaldavReportMethod.java
@@ -577,7 +577,7 @@ public class CaldavReportMethod extends ReportMethod {
     /** Get the content now to set up length, type etc.
      */
     Reader in = getNsIntf().getContent(node);
-    resp.setContentLength(node.getContentLen());
+    resp.setContentLength((int)node.getContentLen());
     if (in == null) {
       if (debug) {
         debugMsg("status: " + HttpServletResponse.SC_NO_CONTENT);
